+++ src/gui/windows/quit_confirm.h
@@ -1,0 +1,16 @@
+//
+// Created by kangaroo. on 28/07/2026.
+//
+
+#ifndef COREDECK_QUIT_CONFIRM_H
+#define COREDECK_QUIT_CONFIRM_H
+
+#include "../context.h"
+
+namespace CoreDeck {
+    void RequestQuitConfirmation(Context &context);
+
+    void BuildQuitConfirmWindow(Context &context);
+}
+
+#endif // COREDECK_QUIT_CONFIRM_H